Codex P1: floor_div overflow, WASM width truncation, raw-bytes HMAC
floor_div (P1) - floor_div(i64::MIN, -1) overflowed a fixed-width divide and
PANICKED (UB across the C ABI). Now returns the wrapping negation (a / -1 == -a,
floor(-a) == -a), panic-free; floor_mod uses wrapping arithmetic for the same
edge. The bindings reject this one input explicitly (true result 2^63 is
unrepresentable): loom_py raises ValueError, loom_c_abi returns -1. Test added.

WASM width (P1) - loom_wasm took u32 seeds and truncated the u64 dice sum + the
i64 floorDiv to 32 bits, so the SAME seed/input could give a different result in
the browser than in Python/C/Rust. Widened seeds to u64, rollDice to u64,
floorDiv + bandFromDistanceFt to i64 - all bigint in the generated .d.ts now,
matching every other surface. No more boundary truncation.

raw HMAC (P2) - the C-ABI HMAC was C-string-only, so an interior NUL truncated
the message. Added hmac_sha256_hex_bytes to the core and loom_hmac_sha256_hex_raw
(ptr+len) to the C ABI so arbitrary bytes hash faithfully; the C-string form
stays as convenience. Header regenerated.

Verified: cargo test --workspace green (loom_math 4 incl the overflow edge);
wasm-pack build clean (bigint surface); loom_py wheel rebuilt, native parity
25/25, floor_div overflow rejected.

+++ rust/loom_wasm/src/lib.rs
@@ -21,9 +21,11 @@ pub fn version() -> String {
 
 // ---- range bands ----
 
+// Codex P1: feet is i64 to match the core (a JS bigint), so a TS caller and the
+// WASM core classify the SAME integer - no i32 truncation at the boundary.
 #[wasm_bindgen(js_name = bandFromDistanceFt)]
-pub fn band_from_distance_ft(feet: i32) -> String {
-    range_bands::band_from_distance_ft(feet as i64).to_string()
+pub fn band_from_distance_ft(feet: i64) -> String {
+    range_bands::band_from_distance_ft(feet).to_string()
 }
 
 #[wasm_bindgen(js_name = bandWithin)]
@@ -32,27 +34,32 @@ pub fn band_within(band: &str, max_band: &str) -> bool {
 }
 
 // ---- deterministic dice (PCG32) ----
+// Codex P1: seeds are u64 and the dice sum is u64 (JS bigint), matching the core
+// + the Python/C surfaces. A u32-only WASM API truncated both vs every other
+// binding - same seed could give a different result. Fixed by using full width.
 
 #[wasm_bindgen(js_name = rollDie)]
-pub fn roll_die(seed: u32, sides: u32) -> u32 {
-    Pcg32::seeded(seed as u64).roll_die(sides)
+pub fn roll_die(seed: u64, sides: u32) -> u32 {
+    Pcg32::seeded(seed).roll_die(sides)
 }
 
 /// Sum of `count` dice of `sides` faces from a fresh seed (deterministic).
 #[wasm_bindgen(js_name = rollDice)]
-pub fn roll_dice(seed: u32, count: u32, sides: u32) -> u32 {
-    Pcg32::seeded(seed as u64).roll_dice(count, sides) as u32
+pub fn roll_dice(seed: u64, count: u32, sides: u32) -> u64 {
+    Pcg32::seeded(seed).roll_dice(count, sides)
 }
 
 /// The next raw u32 from a seeded PCG32 (for parity-vector checks from JS).
 #[wasm_bindgen(js_name = pcg32Next)]
-pub fn pcg32_next(seed: u32) -> u32 {
-    Pcg32::seeded(seed as u64).next_u32()
+pub fn pcg32_next(seed: u64) -> u32 {
+    Pcg32::seeded(seed).next_u32()
 }
 
 // ---- cross-language integer division contract ----
+// Codex P1: i64 in + out (JS bigint), so floorDiv(i32::MIN, -1) no longer
+// truncates to a wrong i32. floor_div itself is panic-free on the overflow edge.
 
 #[wasm_bindgen(js_name = floorDiv)]
-pub fn floor_div_js(a: i32, b: i32) -> i32 {
-    floor_div(a as i64, b as i64) as i32
+pub fn floor_div_js(a: i64, b: i64) -> i64 {
+    floor_div(a, b)
 }
